chore: move @/server/* to @/server/service/*
